+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
 
 - **[BREAKING CHANGE]**: `Bootstrapable` has been renamed to `Bootstrappable` (fixed typo).
+- Changed return type of `initialize` method in `Bootstrappable` class from `Future<void>` to `FutureOr<void>`.
 - Improved documentation with more realistic usage examples.
 
 ## 0.1.0